pyproject.toml
[project]
name = "omni-llm-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
    "uvicorn>=0.43.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
    "types-python-dateutil",
    "pre-commit>=3.7.0"
]


[tool.ruff]
target-version = "py311"
line-length    = 100
indent-width   = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long — ruff format handles it
]

[tool.ruff.lint.isort]
known-first-party = ["providers", "config", "browser"]

[tool.ruff.format]
quote-style            = "double"
indent-style           = "space"
skip-magic-trailing-comma = false
line-ending            = "auto"